requirements.txt
psutil
plyer
customtkinter
pystray
Pillow
